Restructure project into distinct packages, implement USL service
