+++ bot.py
@@ -1,9 +1,8 @@
 import discord
-from discord.ext import commands, tasks
-import os
+from discord.ext import commands
 import json
 import logging
-import config
+from unity_util import bot_config
 import sys
 
 with open('config.json', 'r') as f:
@@ -11,7 +10,7 @@ with open('config.json', 'r') as f:
 
 logging.basicConfig(level=logging.INFO, format="[%(levelname)s]\t %(name)s: %(message)s", handlers=[
     logging.StreamHandler(sys.stdout),
-    logging.FileHandler(f'./logs/{config.LOGGING_FILENAME}')
+    logging.FileHandler(f'./logs/{bot_config.LOGGING_FILENAME}')
 ])
 
 intents = discord.Intents(
@@ -26,7 +25,7 @@ intents = discord.Intents(
     guild_reactions=True,
     dm_reactions=True
 )
-client = commands.Bot(command_prefix = config.DISCORD_PREFIX, intents=intents)
+client = commands.Bot(command_prefix = bot_config.DISCORD_PREFIX, intents=intents)
 
 @client.event
 async def on_ready():
@@ -37,25 +36,25 @@ async def on_ready():
 @client.command()
 @commands.is_owner()
 async def load(ctx, extension):
-    client.load_extension(f'cogs.{extension}')
+    client.load_extension(f'unity_cogs.{extension}')
     await ctx.send(f'{extension} loaded')
     logging.info(f'{extension} loaded')
 
 @client.command()
 @commands.is_owner()
 async def unload(ctx, extension):
-    client.unload_extension(f'cogs.{extension}')
+    client.unload_extension(f'unity_cogs.{extension}')
     await ctx.send(f'{extension} unloaded')
     logging.info(f'{extension} unloaded')
 
 @client.command()
 @commands.is_owner()
 async def reload(ctx, extension):
-    client.reload_extension(f'cogs.{extension}')
+    client.reload_extension(f'unity_cogs.{extension}')
     await ctx.send('{} reloaded'.format(extension))
     logging.info(f'{extension} reloaded')
 
 for cog in conf['preloaded']:
-    client.load_extension(f'cogs.{cog}')
+    client.load_extension(f'unity_cogs.{cog}')
 
-client.run(config.DISCORD_TOKEN)
+client.run(bot_config.DISCORD_TOKEN)
