+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup
+
+setup(
+    name='unity_discord_bot',
+    packages=['unity_discord_bot'],
+    include_package_data=True,
+    install_requires=[]
+)
